requirements.txt
numpy
opencv-python
torch
tqdm
pyyaml
